Allow RuboCop to update config during tests
Travis is running into errors because the style task runs in the same
context as the specs. The process requires and enables WebMock so no
HTTP requests are allowed while the test suite runs.

However, if RuboCop's config drifts out of date, then it attempts to
request a new configuration file. Since it is in the same context
where network requests are not allowed, it throws an exception.

This moves the RuboCop run into its own process where HTTP requests
are allowed.