Add Hound configuration
Hound lets us set a global `.hound.yml` file for an organization, which
this is intended to be. They look in the root of the project, so I've
put `.hound.yml` there and symlinked it to `ruby/` since that's the only
configuration we intend at the moment. I can see us in the future
needing to split out into a `style-ruby` repo independent of other
stacks, if Hound needs different config files for each. Since their
config supports multiple languages in one file, hopefully/maybe we can
stick to just one.