Increase the nesting allowed in specs
We have a situation like this right now:
```
describe EventCleaner do
  describe '.clean_up_rental_id!' do
    context 'when the rental_id has more than 100 items' do
      context 'when all items are created within the last 3 months' do
      context 'when all items are created more than 3 months ago' do
      end
    end
  end
end
```

We are describing the class being tested, giving context on the method being tested, and then we would like to have two contexts left to organize the specs. The current default is three but we think expanding this to four makes sense to be able to better describe specs.